refactor: mv types
